\name{xlsReadWrite-package}
\alias{xlsReadWrite-package}
\alias{xlsReadWrite}
\docType{package}
\title{Natively read and write Excel files.}

\description{
  \pkg{xlsReadWrite} allows to read and write Excel files natively, 
  without any dependencies such as Excel, ODBC drivers, Perl, Python or Java. 
  Currently supported is the .xls (BIFF8) file format.
}

\details{
  Please check \code{\link{read.xls}}, \code{\link{write.xls}} or the README file 
  for more info.
}

\keyword{package}