\name{xlsReadWrite-package}
\alias{xlsReadWrite-package}
\alias{xlsReadWrite}
\docType{package}
\title{Natively read and write Excel files.}

\description{
  The \pkg{xlsReadWrite} package allows to read and write Excel files natively, 
  without any dependencies such as Excel, ODBC drivers, Perl, Python or Java. Supported
  is the BIFF8 file format (Excel 97-2003).
}

\details{
  Info and downloads:
  \itemize{
    \item README file	
    \item General info: \url{http://www.swissr.org}
    \item Bug tracker: \url{http://redmine.swissr.org}
    \item Downloads from swissr dropbox:\itemize{
    	\item listing: \url{http://dl.dropbox.com/u/2602516/swissrpkg/content}
    	\item individual parts: \kbd{http://dl.dropbox.com/u/2602516/swissrpkg/...} \itemize{
	    	\item \bold{(binary package)} \kbd{...bin/<os>/<majvers>/xlsReadWrite_<pkgvers>.zip}
  	  	\item \bold{(shlib)} \kbd{...bin/<OS>/shlib/xlsReadWrite_<pkgvers>_dll.zip}
    		\item (source package with shlib) \kbd{...bin/<os>/src/xlsReadWrite_<pkgvers>.tar.gz}
    		\item (source package cran version) \kbd{...cran/xlsReadWrite_<pkgvers>.tar.gz.txt}
	    	\item (source package actual version) \kbd{...src/xlsReadWrite_<pkgvers>.tar.gz.txt}
	    }
	    \item (placeholders:	<os>: win32, <majvers>: major R version, <pkgvers>: pkg version,\cr
  	  	every file has a correspondent hash file with same name and '.md5.txt' suffix)
    }
    \item Source code: \url{http://github.com/swissr/xlsreadwrite.git} (checkout README)
  }
}

\keyword{package}