\name{xlsReadWrite-package}
\alias{xlsReadWrite-package}
\alias{xlsReadWrite}
\docType{package}
\title{Read and write Excel files (.xls).}

\description{
  Natively read and write Excel files (.xls/BIFF8) without any
  dependencies such as Excel, ODBC drivers, Perl, Python or Java. 
}

\details{
  See \code{\link{read.xls}} and \code{\link{write.xls}} for the
  commands and the README file or \url{http://www.swissr.org}
  for more info.
}

\keyword{package}
